fix(replays): Fix Replay Console and Dom tabs resetting scroll position on hover
