+++ static/app/views/replays/detail/console/index.tsx
@@ -1,4 +1,4 @@
-import {useRef} from 'react';
+import {memo, useRef} from 'react';
 import {
   AutoSizer,
   CellMeasurer,
@@ -8,12 +8,8 @@ import {
 import styled from '@emotion/styled';
 
 import Placeholder from 'sentry/components/placeholder';
-import {useReplayContext} from 'sentry/components/replays/replayContext';
-import {relativeTimeInMs} from 'sentry/components/replays/utils';
 import {t} from 'sentry/locale';
 import type {BreadcrumbTypeDefault, Crumb} from 'sentry/types/breadcrumbs';
-import {getPrevReplayEvent} from 'sentry/utils/replays/getReplayEvent';
-import useCrumbHandlers from 'sentry/utils/replays/hooks/useCrumbHandlers';
 import ConsoleFilters from 'sentry/views/replays/detail/console/consoleFilters';
 import ConsoleLogRow from 'sentry/views/replays/detail/console/consoleLogRow';
 import useConsoleFilters from 'sentry/views/replays/detail/console/useConsoleFilters';
@@ -27,30 +23,9 @@ interface Props {
 }
 
 function Console({breadcrumbs, startTimestampMs}: Props) {
-  const {currentTime, currentHoverTime} = useReplayContext();
-
   const filterProps = useConsoleFilters({breadcrumbs: breadcrumbs || []});
   const {items, setSearchTerm} = filterProps;
   const clearSearchTerm = () => setSearchTerm('');
-
-  const {handleMouseEnter, handleMouseLeave, handleClick} =
-    useCrumbHandlers(startTimestampMs);
-
-  const current = getPrevReplayEvent({
-    items,
-    targetTimestampMs: startTimestampMs + currentTime,
-    allowEqual: true,
-    allowExact: true,
-  });
-
-  const hovered = currentHoverTime
-    ? getPrevReplayEvent({
-        items,
-        targetTimestampMs: startTimestampMs + currentHoverTime,
-        allowEqual: true,
-        allowExact: true,
-      })
-    : null;
 
   const listRef = useRef<ReactVirtualizedList>(null);
   const {cache} = useVirtualizedList({
@@ -74,15 +49,8 @@ function Console({breadcrumbs, startTimestampMs}: Props) {
         rowIndex={index}
       >
         <ConsoleLogRow
-          hasOccurred={
-            currentTime < relativeTimeInMs(item.timestamp || 0, startTimestampMs)
-          }
-          isCurrent={item.id === current?.id}
-          isHovered={item.id === hovered?.id}
           breadcrumb={item}
-          onClickTimestamp={() => handleClick(item)}
-          onMouseEnter={() => handleMouseEnter(item)}
-          onMouseLeave={() => handleMouseLeave(item)}
+          breadcrumbs={items}
           startTimestampMs={startTimestampMs}
           style={style}
         />
@@ -137,4 +105,4 @@ const ConsoleLogContainer = styled('div')`
   border-radius: ${p => p.theme.borderRadius};
 `;
 
-export default Console;
+export default memo(Console);
